mfd: abc-pcie: avoid potential access past end of callback arrays
handler_payload[] and sys_cb2[] are dimensioned 44 for 32
MSIs plus 12 mux'ed interrupts on MSI 31.  abc_pcie_irq_handler() may
access beyond those bounds if higher-order bits are set in register
SYSREG_FSYS_INTERRUPT beyond the 12 expected interrupts.  Prevent such
access for safety. (The product specification document does not specify
values for unused bits in INTNC so far as I could find.)

Cleanup the symbols related to this: rename MAX_INT (which is too
generic, looks like standard symbol MAXINT, and is inaccurate: it is a
count, not a max value) to ABC_INT_COUNT, rename MAX_ABC_MSI to
ABC_MSI_COUNT for accuracy as well, and add ABC_MSI31_INT_COUNT for the
number of mux'ed interrupts on MSI31 to use in the MSI31 de-mux code.
Compute ABC_INT_COUNT as the total of ABC_MSI_COUNT +
ABC_MSI31_INT_COUNT to make this explicit.

Change-Id: I0fcaf2b6583de8e10742fe94da99af6b7df041ef
Signed-off-by: Todd Poynor <toddpoynor@google.com>