char: oscar: pass PCI BAR ioremap VA to driver via property
Pass the VA of the ioremap'ed TPU memory region in a property, rather
than a abc-pcie driver-specific structure.  The Oscar driver will use
this existing mapping.  If the property is not passed, the oscar driver
remaps the buffer itself based on the IORESOURCE_MEM resource, which
must be passed.

Remove an unused abc-pcie interface for accessing the VA.

Bug: 113307531
Change-Id: I208d2f15f21837538ed501946990f7085af0d627
Signed-off-by: Todd Poynor <toddpoynor@google.com>