defconfig: Enable CONFIG_IDLE_PAGE_TRACKING
This configuration enables the code that allows a tool to mark certain
page frames 'idle'. Which will then trigger the rmap to find out if
those page frames are "really" idle. The interface exposed to user space
is a bit-per-page-frame in /sys/kernel/mm/page_idle/bitmap.

Plan is to study how this feature can be used to study working set of an
Android process over a use case or historically through out its life
cycle. The configuration flag that this CL adds has no 'runtime' impact
unless something in user space explicitly writes to the bitmap.
The only difference as a result will be code added by mm/page_idle.c

Bug: 111694435
Test: Build and boot.

Change-Id: I3c80a7e183d008f26b47acd4cd3266a895f34666
Signed-off-by: Sandeep Patil <sspatil@google.com>