binder: Add instramentation to dump_ref_desc_tree()
b/110201845 and b/30870027 has been seen a few times since code
was added to dump out the ref_desc_tree when we hit the
error condition which appears to be impossible.

In those cases, the rb tree appeared to be correct, so this
didn't provide enough information to diagnose the issue. This
patch adds the missing piece of data that hopefully will indicate
why we break out of the desc assignment loop when we have not
yet found a free desc.

There are two conditions that break out of the loop:

  n==NULL or (ref->data.desc > new_ref->data.desc)

both appear to be impossible, but one of them must
have occurred. Adds "n" and "ref->data.desc" to the
output.

Bug: 110201845
Bug: 30870027
Change-Id: I765060e87111cf747e12b1e51aedfc3c770e44bc
Signed-off-by: Todd Kjos <tkjos@google.com>