smb-lib: Fix incorrect settings of Source pull-up
The variable used in the comparison is wrong. Instead of "power_role"
which is used to switch the CC between the Rp and Rd, the parameter
passed in from the caller that represents the real desired Rp value is
the correct one when choosing the different Rp values.

Bug: 116185426
Test: Rp is switched to Rp-1.5A after PD power negotiation is done.
Change-Id: I4e8bc80f76298941d92a2fa85f9b26bfcbc369ad
Signed-off-by: Kyle Tso <kyletso@google.com>
(cherry picked from commit 977bfdf96d4cdc7249834e7df4c8f3b9de0bab07)