pd_engine: Force disconnection when WLC with internal Vbus on
If the device is Source/UFP, put it on wireless charging will trigger
BUG_ON because extcon_usb and extcon_usb_host co-exist. Fix this by
forcing the device to disconnect when wireless charging is on and Vbus
source is from PMIC. And also update the source capability to 5V/0.5A
in case device enables external Vbus source in later connection.

Bug: 113882222
Bug: 140367237
Change-Id: I4844dd7997b68ae1eb76ec2215fdd6b0ade4745f
Signed-off-by: Kyle Tso <kyletso@google.com>
Signed-off-by: rickyniu <rickyniu@google.com>