usb: pd_engine: implement tcpm_set_current_limit
When device connects to PD charger, pd_engine needs
to set up current value to smb5-lib. We need to set
PD_ACTIVE status because smb5-lib would check it.
max_ma is zero when device connects to a non PD host
via type-A port, so we can set it PD_INACTIVE.

Bug: 112400445
Change-Id: I96e8c6ff4d41133be83df5f66a22a8aafc28b62b
Signed-off-by: pumahsu <pumahsu@google.com>