spi: spi-geni-qcom: use FRAGMENTATION for cs toggle
According to Qcom *CS_TOGGLE flag is used for toggling CS line between
spi words. However, cs_change is used to indicate CS line should be
toggled between messages, as mentioned in the documentation:

(i) If the transfer isn’t the last one in the message, this flag is used
to make the chipselect briefly go inactive in the middle of the message.
Toggling chipselect in this way may be needed to terminate a chip
command, letting a single spi_message perform all of group of chip
transactions together.

So, in order to have that behavior use FRAGMENTATION instead of
*CS_TOGGLE

Bug: 138922965
Change-Id: I88bcfcada00e22d7e00b8e06ae568c189cf15321
Signed-off-by: Fernando Lugo <flugo@google.com>