floral_defconfig: CONFIG_CNSS_GENL=y
This commit enables the inclusion of the cnss_genl driver in kernel
build. cnss_genl driver creates a netlink family and multicast groups
to facilitate communication between WLAN driver and userspace.

The use of this driver avoids the need for net-admin capability for the
userspace processes to communicate with the driver.

Bug: 111143438
Test: Manual test: make sure that cnss-daemon can communicate with the
      driver with no net-admin

Change-Id: I8b1cbc003bcab44b3d3ef19cae1e9a84ffdd81c3
Signed-off-by: Ahmed ElArabawy <arabawy@google.com>