smb-lib: Set Isrc to 180uA in DRP mode
If Isrc is set to 80uA in DRP mode, device will periodically trigger CC
interrupts when attaching to a capative cable with CC pulled-up to Rp-3A
and with a capacitor on Vbus path. This is because voltage division on
the CC pin will fall within the range of vRd-def when CC termination
toggles to Current Source.

Set the Isrc to 180uA (equivalent to Rp-1.5A) will prevent this issue
because CC voltage will be beyond vOPEN.

Bug: 117741987
Change-Id: I2a7ad47ff8ed691c76effd314ab5bd336fb39a46
Signed-off-by: Kyle Tso <kyletso@google.com>
(cherry picked from commit d1d43b970ce4e24fe9b4e6c75d35259b51062721)