input: touch: fts: timestamp for pointer events
Currently, it is possible to see a stream of input events coming from
the touchscreen that would only include MSC_TIMESTAMP reports in-between
SYN_REPORT. This is wasteful, since the user does not know what actually
happened (except for the time that the irq was raised).

This is happening because such events as status changes are still
communicated via irq from the touch IC to the touch driver. In touch
driver, sync happens regardless of the kind of event that occurred. This
is okay (maybe slightly inefficient), but multiple syncs would not cause
additional reports from the driver to the user, since no data has
changed. Once timestamp was introduced, the state now changes between
these syncs, and the events are going to user.

To get around this, keep track of the instances when pointer events are
received in the driver, and only report timestamp in those situations.

Bug: 115942683
Test: manual testing:
1) open getevent -lt
2) touch the screen with multiple fingers, ensure that there aren't any
events that only contain MSC_TIMESTAMP. At the same time, open the
"dmesg -w" and look for calibration events from touch driver. You can
force these calibrations by either hovering very near the screen or
doing some very light touches, or by having large contact area (palm,
thumb) with the screen. Whenever these calibrations occur, the touch
IC would send the status event to the touch driver via irq. That used to
trigger the standalone MSC_TIMESTAMP messages.

Change-Id: I0fbe2d8b91c1bb22b367da38d677b6100c2233ec
Signed-Off-By: Siarhei Vishniakou <svv@google.com>