UPSTREAM: coresight: etb10: Refactor etb_drvdata::mode handling
This patch moves the etb_drvdata::mode from a locat_t to a simple u32,
as it is for the ETF and ETR drivers.  This streamlines the code and adds
commonality with the other drivers when dealing with similar operations.

Signed-off-by: Mathieu Poirier <mathieu.poirier@linaro.org>
Reviewed-by: Suzuki K Poulose <suzuki.poulose@arm.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

(Upstream commit d43b8ec599f90c1f07b1bdd29b0c4b6306726ef2).

Bug: 145687602
Change-Id: I8caccb9c2f9055415958aa3096ed3906319a8075
Signed-off-by: Yabin Cui <yabinc@google.com>