misc: ipu: add debugfs support for min core count
This change adds a debugfs file at this location:

/d/paintbox/power/min_core_enable

Writing a value between 0 and 14 will set the minimum number of
active IPU cores.  The cores will remain active until the value
is changed to a lower number.  If an application is run while
this is set then it will be able to increase the amount of cores
active while the app is running but it will go back to the floor
specified in the debugfs file.

Note that if you change the value while an app is running to less
then the what the app requires you may cause the app's job to
fail.

Bug: 114690381
Change-Id: I52871e43b0cc88edba9360cea2f72197d22b44ac
Signed-off-by: Adam Hampson <ahampson@google.com>