FROMLIST: dt-bindings: connector: Add self-powered property
From USB_PD_R3_0
7.1.5 Response to Hard Resets
Device operation during and after a Hard Reset is defined as follows:
Self-powered devices Should Not disconnect from USB during a Hard Reset
(see Section 9.1.2).
Bus powered devices will disconnect from USB during a Hard Reset due to the
loss of their power source.

Therefore it is necessary to know whether the port belongs to
a device which is self powered or bus powered. This change
adds "self-powered" flag to the connector class which present indicates
that the port belongs to a device that is self powered. Else it is
bus powered usb device.

Bug: 117227841
Change-Id: I921ddf8ee85cc6136f1e45cb2d41bbd70ecd8e6e
Signed-off-by: Badhri Jagan Sridharan <badhri@google.com>