misc: ipu: add debugfs control of jqs logging
1.  Add DebugFS control of JQS log level and sink
The following DebugFS files have been added:

/d/paintbox/jqs/log_level

This controls that the log level that JQS will report at. Setting
the log level will cause messages at that level and lower to be
reported:

Valid values for this file are:
0 - No logging
1 - Fatal
2 - Error
3 - Warning
4 - Info

/d/paintbox/jqs/log_trigger_level

This controls that the log level that JQS will signal the AP at.
Setting the log level will cause messages at that level and lower
to be reported with a signal to the AP.  Note this setting only
takes effect if kernel_log_en is also set.

Valid values for this file are:
0 - No logging
1 - Fatal
2 - Error
3 - Warning
4 - Info

/d/paintbox/jqs/kernel_log_en

This controls whether or not JQS log messages are sent to the
kernel to be emitted to the dmesg log.

Value vlues for this file are:
0 - Kernel logging is disabled
1 - Kernel logging is enabled

/d/paintbox/jqs/uart_log_en

This controls whether or not JQS log messages are sent to the
JQS UART.

Value vlues for this file are:
0 - UART logging is disabled
1 - UART logging is enabled

/d/paintbox/jqs/uart_baud

This controls the baud rate of the JQS UART.  The value can be
set to any integer.  It is up to the user to ensure that the
value is a valid baud rate for the UART.

2.  Consolidate JQS firmware state in a single struct
3.  Add locking for JQS management.

Bug: 116164081
Change-Id: I0f11e550b84ffce9301f7cd25238e7c39e440d1b
Signed-off-by: Adam Hampson <ahampson@google.com>