pd_engine: Don't block psy_changed if typec_mode is NONE
CC-change notification will be blocked in psy_changed_handler if the
psy_changed event of typec-attach-detach irq is executed earlier than
that of cc-state-change irq. The race condition will cause the update
of the OTG and the exit of host mode to be skipped.

------------------------------------------------------------
-> cc-state-change irq
                            -> typec-attach-detach irq
                            -> pe_start set to 0
                            -> typec-attach-detach
                                 psy_changed (pe_start == 0)
                            -> blocked in pd_engine
-> cc-state-change
     psy_changed (pe_start == 0)
-> blocked in pd_engine
------------------------------------------------------------

Fix this by letting the psy_changed go ahead if typec_mode is NONE.

Bug: 117260520
Change-Id: I0403ff0b36d948e19b90e42038fcc4184561f9c7
Signed-off-by: Kyle Tso <kyletso@google.com>