ANDROID: modpost: validate symbol name in find_elf_symbol
If an ARM mapping symbol shares an address with a valid symbol,
find_elf_symbol can currently return the mapping symbol, because
it fails to validate the symbol before returning it. This can
result into bogus warnings:

  WARNING: vmlinux.o(.text+0x18f4028): Section mismatch in
  reference from the function set_reset_devices() to the
  variable .init.text:$x.0
  The function set_reset_devices() references
  the variable __init $x.0.
  This is often because set_reset_devices lacks a __init
  annotation or the annotation of $x.0 is wrong.

This commit adds is_valid_name to find_elf_symbol, similarly to
how it's used in find_elf_symbol2.

Bug: 117237524
Change-Id: I7bcab44f7c8f350baa699cd966c04cfa5f81ca0a
Signed-off-by: Sami Tolvanen <samitolvanen@google.com>