arm64/dts: fix black screen when AOD out for s6e3fa7 panel
Currently we have two stages for AOD in command. It will disable
display in 1st stage and then enable display in 2nd stage. If
we press power key to force to abort the sequence between these
two stages, it will disable display and then execute AOD out
command, the display will not be enabled. Add display off and
on command for AOD out to fix this bug.

Bug: 117398219
Test: no black screen in the stress behavior
Change-Id: I1b968d121fcfd4e0a4bee4cdbde84c18ad1e6a5a
Signed-off-by: Chris Lu <luchris@google.com>