input: touch: heatmap: Properly lock heatmap buffers
The access (r/w) to the list of heatmap buffers was protected by the
heatmap lock. However, the actual usage of the buffers was not.
Therefore, once a buffer is dequeued from the list, it was being
modified without the lock protection. This lead to the following
problem: during teardown, if a touch interrupt comes in, the buffers can
be torn down but still be accessed from the heatmap read. This caused
kernel panic (leading to ramdump).
To reproduce the bug, write a program that accesses the heatmap video
device, and then crashes (happened during HAL development). Next,
continuosly touch the screen (to keep generating interrupts), and launch
the HAL process. The phone will enter ramdump after HAL crashes.

To fix this, move the bus read into a separate step in the interrupt
routine. Now there's a private temporary buffer on the stack that the
bus read is pulling data into. Once the bus read is finished, the data
from the temporary buffer gets copied into the actual (user-readable)
heatmap frame in a single locking operation.

This change means that we no longer (*) check if there are available
buffers before we begin the bus read. This could lead to inefficiency,
since at the end of the bus read, we may find out that there aren't any
free buffers, so we will drop the frame. That would mean that the bus
read was wasteful. In real life testing, however, such a situation was
not observed.

* we do actually check
Still, to account for this possibility, we add an optimization. If we
see that this situation happens at least 3 times in a row, we will start
checking for the availability of free buffers before doing the bus read.
This would also help prevent logspam, which currently happens in a
situation where the user process decided to stop reading buffers, but
did not stop streaming.

We don't want to unconditionally check for availability of free buffers
before the bus read, because it is possible that a buffer may become
available during the read itself. The read can take several
milliseconds, so this could give the user space additional time to
process the existing data and return a buffer.

To check for whether the streaming is currently on, use official api
"vb2_is_streaming", instead of using hidden api "streaming".

Small additional fixes:
- Ensure consistent use of strength_t
- Add a doc for heatmap_read

Bug: 117469302
Test: crashes no longer observed using the steps described above
Change-Id: I2b77a79469e22ce4ee39827a951a2ace58db8972
Signed-Off-By: Siarhei Vishniakou <svv@google.com>

(cherry picked from commit b5a5623e4d8ef20afa04c7d28d32413a911f8677)