arm64/dts: fix brightness flicker when exiting LP mode for sofef00 panel
The brightness value 51h can not be zero in LP mode for sofef00. So we
set it to max in current LP command. But the existing brightness command
will cause sudden high brightness when exiting LP mode because DDIC will
use 51h value immediately when it changes to normal mode. Remove the max
brightness and add 51h commands in LP mode properties. It will use
appropriate brightness values when exiting LP mode to avoid brightness
flicker.

Bug: 117588962
Test: check brightness when exiting LP mode
Change-Id: Ia49d83ea20d41ff99982a2cdb4bcb19d69431ca2
Signed-off-by: Chris Lu <luchris@google.com>