drm/msm: dsi-staging: fix parsing of command line override
The current parsing is broken when there are multiple config keywords
in command line. The kstrtol API only reads numbers if there's a
terminating nul character, but when there are multiple keywords it may
reach a colon. Use sscanf instead to read up to the numeric value an
ignore the rest.

Bug: 115369807
Change-Id: I01000110eb773e8f24e1e91eedcd17990e48961b
Signed-off-by: Adrian Salido <salidoa@google.com>