char: oscar: gasket: add interface to set a separate DMA device
In at least the case of an mfd_cell multi-function device child, the
child platform device does not inherit the DMA / IOMMU configuration
of the parent.  Copying that configuration is non-straightforward or
even likely impossible when the parent device is PCI and holds
pointers to IOMMU-related state only made available for PCI devices.

Add a gasket call to allow platform chip drivers to register their
PCI parent as the device to use for DMA API calls, such that the
chip drivers can workaround this.

Bug: 113307531
Change-Id: I36c489e4fe8e9f5900b5de82cec7c00de603bc68
Signed-off-by: Todd Poynor <toddpoynor@google.com>