regulator: tps: Add support for TPS regulator
This driver adds support for the TPS fixed voltage regulators
wired to a set of GPIOs and when the GPIOs are needed to be flipped
on/off in a specific sequence to enable/disable regulator.
The driver is needed when more than one gpio pin is needed to control
the regulator. Optionally, a delay in microseconds can also be
specified after enabling/disabling each gpio.

Bug: 77536221
Change-Id: I17a5e501a526653116239c169bd86aab895b9171
Signed-off-by: Badhri Jagan Sridharan <badhri@google.com>