input: touchscreen: stm: rework interrupt control
The driver counts the number of calls to disable_irq and attempts
to repeat that number of calls to enable_irq when interrupts are
enabled. This logic is unnecessary, complicated, and still managed to
allow an "Unbalanced enable for IRQ 237" error during boot. This patch
replaces the disable-counting logic with a simple state variable.

Bug: 118706127
Change-Id: I8b8f2bac12790383d56ef7f2f1fd1ee32f20cab0
Signed-off-by: Steve Pfetsch <spfetsch@google.com>