power: p9221_charger: delay reschedule of ramp work
The driver can get into a situation where resume has not completed
and there is constant scheduling of the ramp work leading to spamming
of the log due to debug print.

Since resume can take some time after the initial alarm, prevent
constant rescheduling by adding a delay to the reschedule and
also move the debug prints around to minimize the spam.

** 1 printk messages dropped ** [51715.534173] c1  31939 p9221 0-0061: Ramp reschedule

Bug: 120216475
Change-Id: Idca46d7d24c0f033a3d6d93a54608f83232ff289
Signed-off-by: Patrick Tjin <pattjin@google.com>