staging: s2mpg01: Fix driver file paths
s2mpb04 was the code name for previous generation.  The naming should
be corrected to avoid confusion.

Renames directory from drivers/staging/s2mpb04/ to
drivers/staging/s2mpg01/, as well as its files.

Minor changes to source code to meet checkpatch.pl.

Otherwise, source code and device-tree are NOT touched in this patch,
but will be addressed in future patches.

Bug: 118705469
Test: ls /sys/bus/i2c/devices/4-0066/
Change-Id: Ia0212ea96dd66c8d36b0dd72110f8ccbe4f875ab
Signed-off-by: Cheng Gu <gucheng@google.com>