staging: s2mpg01: Fix function and macro naming
s2mpb04 was the code name for previous generation.  The naming should
be corrected to avoid confusion.

Renames all functions and macros:
  S2MPB04 => S2MPG01
  s2mpb04 => s2mpg01

Device-tree related strings are not modified, and will be addressed in
future patches.

Note that there is still mismatch between s2mpg01-core.h and spec.  This
will be addressed in future patches.

Bug: 118705469
Test: ls /sys/bus/i2c/devices/4-0066/
Change-Id: I4a8a1e1973468288160a84b6f2f4516087268e94
Signed-off-by: Cheng Gu <gucheng@google.com>