ANDROID: fix up hwmon Makefile typo
Commit 0165a06db42c ("ANDROID: HACK: Makefile: fix all build warnings
for in tree code") added a "-Wno-no-enum-conversion" warning line to
drivers/hwmon/Makefile.  There is no such warning flag, so drop a "no-".

No idea why this didn't trigger until now, but probably due to a bunch
of clang fixes ending up in the upstream LTS release (4.4.165) which
caused this to suddenly hit.

Bug: 62058353
Cc: Nick Desaulniers <ndesaulniers@google.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>