misc: ipu: separate link pre-down and link failure
The PCIe EP/MFD driver will send a link pre-down notification to
the IPU driver prior to the link going down.  This change fixes
the handling of that scenario.  A link pre-down is an order
shutdown of the link.  The IPU driver will bring the JQS down and
unstage the JQS firmware.  With the link down the driver will need
to restage the firmware when the link is up again.

Link failure continue to be handled in a similar fashion to JQS
watchdog failures.

This also includes some cleanup of the JQS hardware checks and
error reporting.

Change-Id: I86a10bff9587c2d27c418dbb59ad67359f0f8027
Signed-off-by: Adam Hampson <ahampson@google.com>