drm/msm: delay backlight update until the first commit
In the current implementation panel set backlight can be called even before
receiving a frame from the driver during the first commit. Hence there is
chance for some garbage content to be shown onto the panel. So this change
imposes a condition to delay backlight update of the panel until the
first frame is received from the HW.

Bug: 121002488
Test: no corrupted image during off-mode charging
Change-Id: I9c1421d47cb683f74a39dfb2ef09f53f1fdb648f
Signed-off-by: Weizhung Ding <weizhungding@google.com>