mm: refer the backing device for readahead
Current backing device readahead size is stored in file
handle when creating. However, when a backing device's
readahead changes, it would not take effect unless this
file handle is closed and recreated. Thus, if a process
always keep the file handle, it would always use the same
readahead size associate the file. Correctly, We should
use this backing device inode to find the all linking file
handles to renew the value. However, this might take a lot
of effort when changing the value. Instead of traversing
all file handles, use the min value to judget the ra size.

Bug: 122216621
Test: boot
Change-Id: I3c43ccdd10a2e3f3aceebf4798d01f38de902f9f
Signed-off-by: Martin Liu <liumartin@google.com>