char: oscar: gasket add device-managed PTE allocation
Device driver calls to map pages may request that a new driver callback
be invoked to allocate, update, and deallocate extended level-1 page
tables for the mapping.  This is typically useful for co-locating page
tables with buffers that reside on the TPU device when there are
performance advantages to avoiding the host memory access during MMU
lookups.

Actions to map and unmap device-managed PTEs are implemented, assuming
the device-specific memory for these may not be automatically accessible
to the host and that mappings for these consume scarce resources (such
as PCIe BAR mapping windows).  PTEs are allocated and mapped for CPU
access one at a time instead of combining these operations into one
large allocation plus mapping for all PTEs needed for simplicity,
assuming the allocation and mapping operations are relatively
inexpensive.  Initializing unmapped PTE entries to zero/invalid is
optimized to avoid duplicate initialization at allocation time of
entries that will be written later as part of the current mapping
operation, under the assumption that PTE entry writes are expensive
and unnecessary writes should be avoided.

If any parts of the requested TPU virtual address range already have
level-1 PTEs allocated then the existing page tables continue to be
used, whether or not the PTEs are device-managed.  This is also true
for mappings of host buffers where device-managed PTE allocation is not
requested, if a device-managed PTE is found already allocated for any
portion of the mapped TPU address space.  This may have performance
implications, depending on how frequently the mappings for the address
space range are updated (and the cost of mapping the PTEs to the host
CPU and updating contents over PIO) versus how often the mapped buffers
are accessed by the TPU.  It is expected the TPU runtime separates
on-device versus on-host buffer mappings in the TPU virtual addresses
requested to avoid this situation if needed.

Bug: 122369225
Test: RNN-T 107M, spaceparkmobilenet224resdepth100,
      mobilessdfacekeypoints100
Change-Id: I32c2665607629c33ecb63fb4b66a000d40f605cf
Signed-off-by: Todd Poynor <toddpoynor@google.com>