misc: ipu: refactor shared buffer apis
This change makes the memory allocation calls more specific regarding
that a shared buffer specifically is being allocated, synced, or freed.
The mapping and unmapping of the bar has also been moved to the adapter
layer since this is specific to the allocation implementation and not
the act of allocating or freeing a shared buffer.

Bug: 118455152
Change-Id: I072c399a84a93a58c4821b96a039d498824ddc76
Signed-off-by: Adam Hampson <ahampson@google.com>
Signed-off-by: Blair Conner <blairconner@google.com>