touchscreen: sec_ts: Fix potential race in sec_ts command execution code path.
sec_ts driver used a mutex to serialize access to a flag indicating whether any
command is currently running. But some code path reading the status of
the flag were not protected which could potentially result in race. In
any case a mutex for a flag seem an over-kill. This CL got rid of this
mutex by converting the flag to an atomic variable.

Bug: 123027495

Change-Id: I4537241d5d10dc97dd764fe8324afb35c414668c
Signed-off-by: Biswajit Dash <bisdash@google.com>