char: oscar: set gasket hardware status dead if PCIe link is down
Use the gasket hardware status API to set the oscar device hardware
state "dead" while the device is unavailable, typically due to the PCIe
link being down to conserve power.  This tells certain non-performance-
critical paths in gasket, such as shutdown and fd close processing, that
the action being processed has occurred at a time in which the Oscar
device is unavailable and the device must not be accessed.

Bug: 123062257
Change-Id: I8a6a694badf4fbe07bf2c77eb74de2e24336c8be
Signed-off-by: Todd Poynor <toddpoynor@google.com>