msm: camera: Fix sometimes delay to preview on IR Cams
The root casue of this problem is it doesn't receive done irq
normally when cap sense is going to initialize cam cci. Then,
waiting for 1.5 seconds timeout.

To avoid this case, add a protection that reading the mask data
in that address before waiting for irq first. If the got value
is consistent with the expected value to be written, we do not
need to keep waiting for the irq.

Bug: 123063888
Test: Test IR Cams with laser and all camera basic functions
Change-Id: I0c314d8862bb480daa2ee7f25793791ffe734c47
Signed-off-by: pointerkung <pointerkung@google.com>