misc: airbrush: handle suspend during sleep states
During sleep, DDR is kept in self-refresh mode and the MIF input
clock is moved from PLL to Oscillator.
But when Suspend state is requested, as MR registers can't be
configured at Oscillator clock frequency, the MIF input has to be
moved to PLL before configuring MR timing registers for 1866MHz.

Bug: 122900226

Change-Id: I7e86f7476b821cf8aca1ac951a302117546e22c3
Signed-off-by: Shaik Ameer Basha <shaik.ameer@samsung.com>