build.config: Copy kernel image to dist directory
This CL is part of the process to move away from dtb images
appended to the kernel image and instead make them a part
of boot image by modifying the boot image header. The next step
will turn off the CONFIG_BUILD_ARM64_APPENDED_DTB_IMAGE config
and revert the CLs that added the same.

Test: Image.lz4 present in dist directory
Bug: 111136242
Change-Id: Ie32734d7d9f5c286a7ab3127afb9d573a66b8a0e
Signed-off-by: Hridya Valsaraju <hridya@google.com>