misc: ipu: shift rpm work to start/stop hooks
The IPU core driver uses a generic power domain to manage the
JQS power state.  The client driver will use runtime pm to
request the JQS come up and release it to allow it to go to a
low power state.

The original implementation used the gpd power_on/power_off
hooks to control the JQS power state.  The problem is that these
hooks are used for both device power management and runtime power
management.  This change shifts the JQS pm work to the start and
stop hooks which are only called for runtime power management.

The power_on and power_off hooks will be used only for device
power management.  The IPU client takes a wakelock to prevent
suspend while the JQS is operation.  In the event that the JQS
is up when going into suspend (This should not happen) then the
JQS will be powered down and the recovery path invoked on
resume.

Bug: 123233983
Change-Id: Ic7bd756c192514727a437079e30cdd4e65300957
Signed-off-by: Adam Hampson <ahampson@google.com>