l2tp_ppp: Fix Protocol in PPP Frame
protocol-field-compression is supported for PPP Frames. Therefore,
Protocol field has to be fixed before sending it to ppp driver.
Failing to do so, kernel retrieves 2 bytes and incorrectly
assumes that the protocol is incorrect and unsupported.

Check if the protocol is compressed, and fix it by pushing one
additional byte to skb.

CRs-Fixed: 2393404
Change-Id: Ia31806a29e038b5b7d8829498bf42491b94553df
Signed-off-by: Devi Sandeep Endluri V V <dendluri@codeaurora.org>