misc: airbrush: Check bitwise flags in PCIe status notifier
As PCIe link status notifier types have been expanded, it is possible
that PRE_DISABLE and POST_ENABLE may be bitwise OR'ed with other
types.

With this patch, subscribers only check specific bit that they
are interested instead of checking the whole value of the notifier type.
This will avoid missing notifiers.

Bug: 123893256
Test: atest :airbrush-platform-presubmit
Change-Id: Id11368b1238ccc1c290276ba2601549a32024f7e
Signed-off-by: Cheng Gu <gucheng@google.com>