misc: faceauth: calculate copysize correctly
hypx implements a simple userbuffer to scattered-gathered list
converter. For that it allocates physically contiguos regions starting
with a large page order. If memory is fragmented and imposible to get an
allocation of such order then copier falls back to smaller page order.

We need to make sure that number of copied bytes is adjusted for smaller
page order allocation.

Test: boot at P19, run faceauth el2 with faceauth_ctl tool
Bug: 123958559
Change-Id: If9d9874c178c3644dd13f51167240cf63b93c6ca
Signed-off-by: Anatol Pomazau <anatol@google.com>