misc: airbrush: Add adjustable smps2_delay; default to 40 ms
This CL adds 2 adjustable delay at the time of turning off
ldo4 and smps2 rails.

smps2_delay is set to 40 ms by default.
Default value for s60_delay is changed to 0, because smps2_delay
should serve the same purpose when transitioning to S60.

s60_delay can be used for extra delay needed for entering S60.

The default 40 ms of smps2_delay is to mitigate a stability
issue during suspend/resume seen on most EVT devices.

Impact of this change is +40 ms when entering S50.  Other state
transition latency should not be affected.

The provided debugfs entry adds capability to sweep
the delay values in runtime.

Bug: 123771153
Test: atest :airbrush-platform-presubmit
Change-Id: I684b37add0a12a28d72125deece4ac54fc5f2128
Signed-off-by: Cheng Gu <gucheng@google.com>