defconfig: remove CONFIG_RMNET_DATA_DEBUG_PKT on perf_defconfig
According to a part of define of CONFIG_RMNET_DATA_DEBUG_PKT, we have to remove it on perf build.
"Use of this will have major performance impacts and may even trigger watchdog resets if too much is being printed. Hence, this should always be compiled out unless absolutely needed."

Bug: 121199217
Change-Id: I4f362e6692302d168b04ad70f8dd66a25a3a49cd
Signed-off-by: Julius D'souza <jdsouza@google.com>