clk/qcom/mdss: fix flags for DSI and DP PLL clocks
Remove the CLK_GET_RATE_NOCACHE flag from all the DP and DSI
pll clocks. This will eliminate the need to recalculate the
clock rates from HW registers when querying the rates for the
PLL clocks. This will ensure that no unclocked register accesses
are done when these clocks are queried while the display core
is power collapsed.

Bug: 119283453
Change-Id: I6541fa7ac2bc611d0595ba78eb145b3938f5c568
Signed-off-by: Aravind Venkateswaran <aravindh@codeaurora.org>