char: oscar: rename interrupts to identify the interrupt line
Use the internal names of TPU interrupts as passed from the parent mfd
driver as the interrupt names registered with the system, rather than
the device name.  This allows tracing tools to identify which IRQ number
is which (the numbers are allocated dynamically as PCIe MSIs are mapped
to unused GIC IRQs on floral).

Bug: 123237665
Change-Id: I0d00f00efc5ece7fd8b9aca0dbacd347f2c35ade
Signed-off-by: Todd Poynor <toddpoynor@google.com>