misc: ipu: added feature to track runtime requests
This adds a boolean variable to track when the JQS runtime has
been requested to start or stop.  Based on this boolean the IPU
driver will enabled or disable the JQS firmware when clock (IPU
or DRAM) change notifications are received.

This change also refactors the bus notification management code
so it is concentrated in ipu-core.c.  An attempt to resume the
JQS will be made whenever a clock is enabled.  The JQS will be
resumed if the JQS has been requested by the client and all
hardware dependencies have been met.

Bug: 124020192
Bug: 123312891
Change-Id: I3af68dac149b4d5b40570a0c5cad0003fc508cfd
Signed-off-by: Blair Conner <blairconner@google.com>
Signed-off-by: Adam Hampson <ahampson@google.com>