display: add the power handling of boost IC
Pulling down AVDD_EN (GPIO#12 on PM660) is to enable power saving
for the boot IC.

The power saving will be enabled if panel is in idle/off state.
Disable the power saving otherwise.

In such way, power consumption could be reduced to a certain extend.

Also, note that this change depends on hardware schematic/component of
display.

Definitely not for all in-cell panel solution.

BUG: 124376266

Change-Id: Ie996d3bb7c8c07ccbcf9906aa23c641b2fec1071
Signed-off-by: chester_sun <Chester_Sun@compal.com>