usb: typec: tps6598x: Check mode of operation
To prevent loading of the driver when the PD controller is
still in some operational mode that the driver does not
support, checking the mode in driver probe callback
function.

TI PD controllers may be in undefined mode of operation
for example when the application code (firmware) is
completely missing.

Signed-off-by: Heikki Krogerus <heikki.krogerus@linux.intel.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

Bug: 143059282
Change-Id: I9eb457dacbb3dc19e054d39884141d6f9dd79cb4
Signed-off-by: Kyle Tso <kyletso@google.com>