misc: faceauth: Set RWX for pages returned from EXT_DSP to HLOS
Jawad saw a crash with following EL2 message:

 Core:7, Instruction abort from EL0, Permission fault 2nd level

 Faulting Address: 0x11e47e000
 Address of Last Instruction Executed: 0x5ff8105b30 at EL0
 System Registers:
  ESR_EL2(8200000e)
  FAR_EL2(5ff8105b30)
  HPFAR_EL2(0)
  ELR_EL2(5ff8105b30)
  HCR_EL2(80080005)

It says that userspace is trying to execute an instruction and got a
permission fault (non-executable page) from the second-level mapping.
i.e. first level (kernel) mapping is fine and has the executable bit, but
EL2 does not.

One way it can happen is that memory returned to HLOS been
hyp_assign'ed without exec bit.

To fix the problem we need to use the EXEC flag.

Bug: 124399711
Test: boot at P19 and run stress tests
Change-Id: Ie8c94abf8704f92be6875860e0c2e17c931b3258
Signed-off-by: Anatol Pomazau <anatol@google.com>