misc: ll-pool:add linked list based remote allocator
Creating a memory allocator based on linked list (ll-pool) to
support non-contiguous allocation for remote memory. It is to
be used in ab-dram component to replace current genpool allocator.

The reason this component is needed is that the genpool allocator
does not allow non-contiguous allocation, and there is no off-the-shelf
solution that does that.

Bug:128546831

Change-Id: I8af9a44f8b19ffbbdb8b193f64faa6e65a3f9d2a
Signed-off-by: Alex Tang <pztang@google.com>