usb: typec: tcpm: AMS and Collision Avoidance
This patch provides the implementation of Collision Avoidance introduced
in PD3.0. The start of each Atomic Message Sequence (AMS) initiated by
the port will be denied if the current AMS is not interruptible. The
Source port will set the CC to SinkTxNG if it is going to initiate an
AMS, and SinkTxOk otherwise. Meanwhile, any AMS initiated by a Sink port
will be denied in TCPM if the port partner (Source) sets SinkTxNG except
for HARD_RESET and SOFT_RESET.

Conflict for merge "tcpm: Do not renogtiate unless PD is active" before

Bug: 120257460
Bug: 149184221
Change-Id: Ifef0f4d0c2e2e502fb867fdbd81cfdaa8f3c4c70
Signed-off-by: Kyle Tso <kyletso@google.com>
Signed-off-by: rickyniu <rickyniu@google.com>