display: add the power handling of boost IC
Pulling down AVDD_EN (GPIO#12 on PM660) is to enable power saving
for the boot IC.

The power saving will be enabled if panel is in idle/off state.
Disable the power saving otherwise.

In such way, power consumption could be reduced to a certain extend.

Also, note that this change depends on hardware schematic/component of
display.

Definitely not for all in-cell panel solution.

Change-Id: I65d3fa8d79ccfe55bda99d8b92a21c8a5d0ac9aa
Signed-off-by: william_liu <william_liu@compal.com>