misc: faceauth: Make certain debug call ignore PCIe status
Every time we do a faceauth_process, we will save some debug data
in the kernel ring buffer. Current implementation will stop user
from assessing the kernel buffer when AB PCIe link is down, it's
overkill. With this change, user would be able to assess data
in ring buffer when AB PCIe link is down, dump from AB DRAM is
still rejected under this scenerio.

Test: Boot P19, run atest, then try to shutdown the link from AB side
and see if it works.

Bug: 128531009

Change-Id: I66ca7224a6ba6259618c84fd196a51d412bfb99d
Signed-off-by: Wei Miao <miaowei@google.com>