misc: ipu: removed jqs ipu clock rate setting during firmware disable
The ipu_core_jqs_disable_firmware erroneously set the stored IPU clock
rate to 0. The IPU driver does not have control of the IPU clock rate
and only stores a copy of the reported value from the Airbrush state
manager.

Setting the stored value to 0 caused a problem when the runtime
requested to the JQS to stop and then start again before any hardware
changes were reported from the Airbrush state manager.

Bug: 128541094
Change-Id: I8856477f29d1c1f76b9406d3816db3033ba48377
Signed-off-by: Blair Conner <blairconner@google.com>