battery: fix htc_battery_probe abort in charger mode
htc_battery_probe_process() which got batt_id fell into
EPROBE_DEFER when profile_loaded had not completed.
However, the re-probe did not happen in charger mode and
the probe init was not done.

Because batt_id is not used later, set batt_id to default
type if profile is loading, not block battery driver running.

Bug: 127586646
Change-Id: I15960b77db1c85408dfe4df5d2137ef5e9aa7e2e
Signed-off-by: Jack Wu <wjack@google.com>