thermal: Use separate shutdown thresholds for the thermal core and thermal-engine.
Currently, the thermal core and thermal-engine shutdown threshold are the same.
Need to push the thermal core shutdown threshold to 2 degree C higher for cushion.
Because the thermal core won't provide the shutdown reason but the thermal-engine does,
So we keep the both of them as a backup if the normal mitigation path does not work (thermal-hal).

Bug: 128935264
Test: Verified thermal_core shutdown via emul temp.
[  155.540603] c1   5964 thermal thermal_zone59: critical temperature reached(59 C),shutting down

Change-Id: I87ce723fba77d59d8a9eba9f3b8a07813c906b7d
Signed-off-by: davidchao <davidchao@google.com>